# [generator:region] Speedup locality index building: parallel region object covering