Ускорение генерации regions_index.locidx.
Некоторое регионы требую большого объёма вычислений для покрытия геометрии cell-ами. Такие объекты обрабатываются параллельно.

Сам по себе PR ускоряет генерацию regions_index.locidx менее чем в 2 раза (17m:32s vs 30m:30s), но совместно с PR #79 генерация сокращается в 4 раза (7m:15s vs 30m:30s)